refactor(dashboard): update utils folder name
re #130
